Switch dtype to bfloat16 and make config path optional in argument parser.

+++ config/train_predictor_config.py
@@ -54,8 +54,7 @@ min_lr = 3e-5
 
 # System settings (optimized for RunPod)
 backend = "nccl"
-# dtype = "bfloat16"  # Use bfloat16 for efficiency if available
-dtype = "float32"
+dtype = "bfloat16"  # Use bfloat16 for efficiency if available
 compile = True  # Enable compilation for speed
 keep_alive = False  # Auto-stop by default
 check_nans = False  # Check for NaNs in cloud training
